refactor: extract net worth chart
